Second version of the project.
After the midterm assesment I updated my code based on your comments and based off my own feedback I put at the end of the code. 
This version also has my own feedback of what to improve and what has improved at the end. I will also add this text here below:
Comments on progresss:
1. The use of Selenium resulted in more information being able to get extracted, and automated the browser interaction.

2. I tried other links, the code still worked. This was not the case for the first version of this project. 
In that version the code only worked the specific type of class that I defined in the code. You had
to check the class of the website and change this in the code for the code to be able to function.

3. I added my name, student number, and a name of the project

Things to add/improve:
1. I have troubles with the authors that are linked on the website to another page with their bio. This code 
takes their whole bio and puts it as "Author's Name"

2. I want to add that the code can run on a list of urls. So that when writing a paper this code can be used for creating
the reference page.