Version 3
This is another step in my project. The previous feedback I gave myself have been taken into account and applied in the project. See my feedback on this version below and at the end of the code:

Comments on progresss:
1. I added the option to list the urls and this worked! 

Things to improve/add:
1. there is still trouble with extracting Toby Saunders name proberly. 
His bio gets extracted, not his name...

2. The urls that get extracted take in a lot of space. I want to put them 
in a seperate document on my computer.

3. The plots seem useless. I like it better to have the information in the console.

4. Meet packaging criteria