Final Version of my Project
Comments on changes:
1. I wasn't able to resolve the problem with the linked up author name Toby Saunders.

2. I got rid of the table because I didn't see the use of it. This code is useful
when writing a paper so you can cite faster and also get all the URLS from the website.
This made the code shorter, but the output much clearer

3. When running this code on google scholar articles, it has troubles extracting authors.

This likely happens because I added a function on my laptop that through the TU delft I have access to more studies.
Everytime I open an article I need to log in with my username and password, this confuses the code.

Luckily this code isn't needed for scholar articles given that a lot of times the articles
are provided with a citation of the article itself.  '


4. I put all extrated links in a document on my laptop because it was taking in a lot of space in the 
output making the output messy and difficult to read wwhich goes against the purpose of this project.  


Final thoughts:
This final version of my project defenilty has the cleanest and most usefull output. 
The the addition of the use of selenium in version 2 really upgraded this code significanly 
and made the extraction much easier and convinient. Adding the option to define the urls 
as a list made this code much more applicable for its purpose which is research and writing papers. 

Initially in my proposal I had the idea to also add to extract the numerical data in each url. 
This was removed from the end product because I didn't see much value in this addition. 
Also it was hard to define "numerical data". Numercical data are tables and graphs but 
also numbers in text which was hard to define and extract. Being able to get an advanced 
code that is able to do this reliably would be a way to advance this code.

Lastly I want to add that it lacks the TestPyPi. I made a setup file and a folder which contained all information, however wasn't able to run the TestPyPi in time because had troubles verifying my account.